Issue 10 - Ananke should allow non integers for metric values

+++ src/main/java/com/netuitive/ananke/statsd/client/request/AbstractMetricRequest.java
@@ -12,7 +12,7 @@ import lombok.Data;
 public abstract class AbstractMetricRequest {
 
     String metric;
-    Long value;
+    Double value;
     List<Tag> tags;
     Long sampleRate;
 }
